[DataGrid] Introduce GridRequestQueue wrapper around RequestQueue
